Preserving active tab for cluster page in URL path
